code reconstruction
