+++ dbproxy/dbproxy.go
@@ -1,41 +1,9 @@
 package dbproxy
 
 import (
-    "os"
-    "errors"
-    "io/ioutil"
     sql "elibot-apiserver/sqlitedb"
     Log "elibot-apiserver/log"
 )
-
-var DBName string
-var BackupPath string
-func DBsetup(dbname string, backuppath string) {
-    DBName = dbname
-    BackupPath = backuppath
-}
-
-func RegisterAndQuery(sm sql.SqlMapper, mode int, vars map[string]interface{}) (res string, err error) {
-    err = sm.RegisterSqlMapper(mode)
-    if err!=nil {
-        res = ""
-        return
-    }
-
-    if vars == nil {
-        res, err = sql.Db_query(sm.GetID(), DBName)
-    } else {
-        res, err = sql.Db_query_with_params(sm.GetID(), DBName, vars)
-    }
-    if err!=nil {
-        Log.Error("query fails: ", err)
-        return
-    }
-
-    Log.Debug(res)
-    Log.Print("get_all_metadatas OK")
-    return
-}
 
 func Get_ALL_Arc() (string, error) {
     Log.Debug("in Get_ALL_Arc")
@@ -142,55 +110,4 @@ func Get_All_Zeropoints() (string, error) {
     
     sm := new(sql.ZeroPointSqlMapper)
     return RegisterAndQuery(sm, sql.ELIBOT_GET_ALL_PARAMS, nil)
-}
-
-
-/* For sqlitedb backup, restore and upgrade*/
-func DBBackup() error {
-    Log.Debug("in DBBackup")
- 
-    mgr, err := sql.NewDBManager(DBName, BackupPath, "", sql.DB_BACKUP, 0)
-    if err != nil {
-        return err
-    }
-    return mgr.Exec()
-}
-
-func DBList() ([]string, error) {
-    Log.Debug("in DBList")
-
-    files, err := ioutil.ReadDir(BackupPath)
-    if err != nil {
-        return nil, err
-    }
-
-    var list []string
-    for _, f := range files {
-        list = append(list, f.Name())
-    }
-    return list, nil
-}
-
-func DBDel(Name string) error {
-    Log.Debug("in DBDel")
-    if Name == "" {
-        return errors.New("file name is empty")
-    }
-
-    filename := BackupPath + Name
-    if _, err := os.Stat(filename); os.IsNotExist(err) {
-       return errors.New("file does not exist")
-    }
-
-    return os.Remove(filename)
-}
-
-func DBRestore(BackupName string) error{
-    Log.Debug("in DBRestore")
- 
-    mgr, err := sql.NewDBManager(DBName, BackupPath, BackupPath+BackupName, sql.DB_RESTORE, 1)
-    if err != nil {
-        return err
-    }
-    return mgr.Exec()
 }
